Add edit of groups
